framerate修正

+++ Engine.py
@@ -1,3 +1,3 @@
 class Engine:
     FPS = 60
-    FRAME_RATE = 1000 / FPS
+    FRAME_RATE = int(1000 / FPS)
